[#145005133] Adds a link on the user dropdown below the normal help link. That will take the user to the PDF of the help documentation. They can save that if they want to.
The link expects the help pdf to be titled CPT_User_Guide.pdf. I don't currently have a user guide so the link will throw an error at present. We can add a check to see if the file is there and hide it if there isn't one.

I'm also not sure this will be automatically loaded with the other help files at this time.